requirements.txt
langchain
openai
faiss-cpu
streamlit
streamlit-chat
chromadb
pypdf
Flask
flask-cors